introduce DefaultThemesWithCustomTabsProvider component
